requirements.txt
boto3
botocore
Flask
Pillow
python-dotenv